+++ test/api_utils.test.ts
@@ -373,6 +373,25 @@ describe('gdal_utils', () => {
       gdal.vsimem.release(tempFile)
     })
 
+    it('should reject on exception in the supplied function', () => {
+      const tempFile = `/vsimem/invalid_calc_${String(Math.random()).substring(2)}.tiff`
+      const T2m = gdal.open(path.resolve(__dirname, 'data','AROME_T2m_10.tiff'))
+      const D2m = gdal.open(path.resolve(__dirname, 'data','AROME_D2m_10.tiff'))
+      const size = T2m.rasterSize
+      const throwFn = () => {
+        throw new Error('propagate this')
+      }
+      return assert.isRejected(
+        gdal.calcAsync({
+          A: T2m.bands.get(1),
+          B: D2m.bands.get(1)
+        },
+        gdal.open(tempFile, 'w', 'GTiff', size.x, size.y, 1, gdal.GDT_Float64).bands.get(1),
+        throwFn),
+        /propagate this/
+      )
+    })
+
     it('should reject when raster sizes do not match', () => {
       const tempFile = `/vsimem/invalid_calc_${String(Math.random()).substring(2)}.tiff`
       const espyFn = (t: number, td: number) => 125 * (t - td)